4.21

+++ part4/4.1-4.23/controllers/blogs.js
@@ -27,8 +27,16 @@ blogRouter.post("/", async (request, response) => {
 });
 blogRouter.delete("/:id",async(request,response)=>{
   const id = request.params.id;
-  await Blog.findByIdAndRemove(id);
-  response.status(204).end();
+  const decodedToken=jwt.verify(request.token,process.env.SECRET);
+  if(!(request.token&&decodedToken.id)){
+    return response.status(401).json({error:'token missing or invalid'});
+  }
+  const blog=await Blog.findByIdAndRemove(id);
+  if(blog.user.toString()===decodedToken.id.toString()){
+   return response.status(204).end();
+  }
+  return response.status(401).json({error:"that blog doesn't belong to you :)"});
+  
 });
 blogRouter.put("/:id",async(request,response)=>{
   const id=request.params.id;
